requirements.txt
websocket-client
spotipy
